Mostra os tempos corretamente

+++ trabalho2/src/imageprocessing.h
@@ -11,10 +11,10 @@ typedef struct {
 imagem abrir_imagem(char *nome_do_arquivo);
 void salvar_imagem(char *nome_do_arquivo, imagem *I);
 void liberar_imagem(imagem *i);
-int brilho_colunas(imagem *I, float n);
-int brilho_linhas(imagem *I, float n);
-int brilho_multithreads(imagem *I, float n, int num_threads);
-int brilho_multiprocessos(imagem *I, float n, int num_processos);
+float brilho_colunas(imagem *I, float n);
+float brilho_linhas(imagem *I, float n);
+float brilho_multithreads(imagem *I, float n, int num_threads);
+float brilho_multiprocessos(imagem *I, float n, int num_processos);
 void valor_maximo(imagem *I);
 void teste(char *nome_do_arquivo, int n_de_testes);
 #endif
